+++ Changelog.md
@@ -6,6 +6,12 @@ Bug fixes
 * Fix `change` matcher so that it dups strings in order to handle
   mutated strings (Myron Marston).
 
+Enhancements
+
+* Add support for percent deltas to `be_within` matcher:
+  `expect(value).to be_within(10).percent_of(expected)`
+  (Myron Marston).
+
 ### 2.12.1 / 2012-12-15
 [full changelog](http://github.com/rspec/rspec-expectations/compare/v2.12.0...v2.12.1)